refactor: move component-specific function
