+++ src/features/visual-create/components/create-button.tsx
@@ -1,12 +1,21 @@
 import React from 'react';
 import { Button } from '@fluentui/react-components';
 import { shallow } from 'zustand/shallow';
+import { IAceEditor } from 'react-ace/lib/types';
 
 import { getI18nValue } from '../../i18n';
-import store from '../../../store';
+import store, { getState } from '../../../store';
 import { logDebug, logRender } from '../../logging';
-import { createFromTemplate } from '../logic';
 import { useJsonEditorContext } from '../../json-editor';
+import {
+    IDenebTemplateAllocationComponents,
+    UsermetaTemplate
+} from '@deneb-viz/core-dependencies';
+import { getTemplateReplacedForDataset } from '@deneb-viz/template';
+import {
+    resolveObjectProperties,
+    updateObjectProperties
+} from '../../../core/utils/properties';
 
 /**
  * Displays the content for creating a specification using the selected
@@ -25,7 +34,7 @@ export const CreateButton: React.FC = () => {
     const { spec, config } = useJsonEditorContext();
     const onCreate = () => {
         logDebug('Creating from template...');
-        createFromTemplate(
+        handleCreateFromTemplate(
             metadata,
             candidates,
             spec?.current.editor,
@@ -42,4 +51,70 @@ export const CreateButton: React.FC = () => {
             {getI18nValue('Button_Create')}
         </Button>
     );
+};
+
+/**
+ * For the supplied provider and specification template, add this to the visual and persist to properties, ready for
+ * subsequent editing.
+ */
+const handleCreateFromTemplate = (
+    metadata: UsermetaTemplate,
+    candidates: IDenebTemplateAllocationComponents,
+    specEditor: IAceEditor,
+    configEditor: IAceEditor
+) => {
+    logDebug('createFromTemplate', { metadata, candidates });
+    const {
+        create: { createFromTemplate }
+    } = getState();
+    const jsonSpec = getTemplateReplacedForDataset(
+        candidates.spec,
+        metadata.dataset
+    );
+    const jsonConfig = candidates.config;
+    logDebug('createFromTemplate - processed candidates', {
+        jsonSpec,
+        jsonConfig
+    });
+    const { renewEditorFieldsInUse } = getState();
+    updateObjectProperties(
+        resolveObjectProperties([
+            {
+                objectName: 'vega',
+                properties: [
+                    { name: 'provider', value: metadata.deneb.provider },
+                    { name: 'jsonSpec', value: jsonSpec },
+                    { name: 'jsonConfig', value: jsonConfig },
+                    { name: 'isNewDialogOpen', value: false },
+                    {
+                        name: 'enableTooltips',
+                        value: metadata.interactivity?.tooltip || false
+                    },
+                    {
+                        name: 'enableContextMenu',
+                        value: metadata.interactivity?.contextMenu || false
+                    },
+                    {
+                        name: 'enableHighlight',
+                        value: metadata.interactivity?.highlight || false
+                    },
+                    {
+                        name: 'enableSelection',
+                        value: metadata.interactivity?.selection || false
+                    },
+                    {
+                        name: 'selectionMaxDataPoints',
+                        value: metadata.interactivity?.dataPointLimit || 0
+                    }
+                ]
+            }
+        ])
+    );
+    renewEditorFieldsInUse(specEditor);
+    specEditor.setValue(jsonSpec);
+    specEditor.moveCursorToPosition({ row: 0, column: 0 });
+    configEditor.setValue(jsonConfig);
+    configEditor.moveCursorToPosition({ row: 0, column: 0 });
+    specEditor.focus();
+    createFromTemplate();
 };
